requirements.txt
openpyxl>=3.1.5
customtkinter>=5.2.2
firebirdsql>=1.3.6
passlib>=1.7.4
pyinstaller>=6.14.1
Pillow>=11.3.0
xlrd>=2.0.1
xlwt>=1.3.0